#33 move room message into message folder
