+++ client/sampleLoginClient.js
@@ -6,7 +6,7 @@
 const debug = require('debug')('debug');
 const { connect, emitToServer } = require('./networkUtils');
 const constants = require('../common/constants');
-const { getLogin } = require('./roomMessages');
+const { getLogin } = require('./message/room');
 
 const connection = connect();
